Improved the scaling of the log viewer module.
The log viewer will estimate the number of entries in a log and
decide on one of two view actions.

If the log entries are over 1000, the log will load the file using
fseek to load bottom up the last 1000 entries. (latest entries on
the top)

If the estimate is under 1000 the log will use the default method
and load the entire file top to bottom (latest entries on the bottom)

The log viewer (view) dislays the estimated number of enteries, if
an long ( > 1000) log is loaded the view displayes a message that only
the lastest 1000 rows are shown.

The estimation function has been added to the file helper, and
does not count top to bottom, but uses a sample of the first x
(default being 100), to determine an average line byte size. Unless
the total size is less than the sample size.