Added Generic Humans TXT File
I find that adding a humans.txt file is a excellent idea, It shows the
people behind the scenes who built this site, and when properly done is
a very good idea.  It is unobtrusive so that only people viewing the
source even see it, and you also gain free links from the humanstxt.org
project when they find one on a site.

Please check out http://humanstxt.org/ for the full scoop on this.

I added a generic one for CI-Bonfire, it will need to be edited.