Added new Permission for Profiler Bar, Fixes #375
Added 028_Migration to add new permission, this will cause issue with JFox's Pull Request, I can rename it or rename his, I didn't
to leave a gap in the migrations while we worked on the pull request.  Also added preg_match check for requested pages with image
extensions which helped reduce the amount of 404 Pages not found, ref issue #417.