Added images controller in ignore_pages, Fixes issue #417
This commit plus the last commit has fixed the image redirects, I was having when using the images
controller to dynamically resize a image.   I've been testing this code before commiting it for
about 2 weeks and it has reduced my pages not found to zero in the logs, and has fixed all issues with
login redirects going to a image instead of a proper redirect page.