Use files in assets folder for bootstrap/modernizr
Updated calls to bootstrap and modernizr to point to assets folder
instead of themes folders and removed the files from the themes.
Also modified references to minified versions of the files to point to
non-minified version, to let assets library handle it based on config
(also fixes an issue where both bootstrap.js and bootstrap.min.js were
loaded if minification was not enabled in some cases).