profiler: cellspacing="0" is invalid and we don't need it
html5.validator.nu: The cellspacing attribute on the table element is
                    obsolete. Use CSS instead.

http://stackoverflow.com/questions/339923/how-to-set-cellpadding-cellspacing-in-csshttp://stackoverflow.com/questions/339923/how-to-set-cellpadding-cellspacing-in-css/3209434#3209434

bootstrap.css already used the technique described on stackoverflow
to ensure 0 cell spacing as a default for all tables.
That's good, because none of the other tables generated
by the profiler were using cellspacing="0".

Although, the profiler can also be shown in the front-end.
So we'd better add the CSS in profiler_template anyway,
to be on the safe side.

Sanity-tested by overlaying before/after screenshots of the affected table.