roles settings: fix permission matrix clickable headers
Detected by HTML5 validator.
(cellIndex is not a valid HTML attribute).

We were also setting it to the wrong values,
so clicking on the header affected the wrong column.

Since we're using jquery .eq() to select the cells in the table column,
let's use jquery .index() to determine the index of our column.