translate: cleanup to not use input->get() and input->post() together
It seems to work fine, but it's surprising to see forms submitted to
urls which have query strings in them.  Let's replace the query string
parameters with normal CodeIgniter parameters.

This means the parameter can't be handled in shared code in the
constructor.  But it was only shared by two methods, and the real
common code was limited to a single line.  Moving code into index()
allows duplication to be reduced, so there's an overall reduction in
lines of code.

This also removes several & characters which had not been correctly
escaped for HTML.