fix three labels supposed to be associated with groups of radio-/checkboxes
(Reminder: <label for="can_delete"> targets the element with _id_
"can_delete", not all the form elements with a _name_ of "can_delete").

Replacing the label with <legend> (and the enclosing div with <fieldset>)
would be best for screen-readers (nested fieldsets are perfectly legal).
However bootstrap's CSS doesn't support that usage.

Fall back to WAI-ARIA.  Tested with Orca Screen Reader.  Both e.g.
"Removable?" and "Yes" are read, but only after disabling CSS -
reported as Orca bug <https://bugzilla.gnome.org/show_bug.cgi?id=680249>.

The behaviour is not well-specified.  We don't want to use role=radiogroup
because it's not compatible with the normal html radio buttons (and yet it
"SHOULD" be present for *all* radio buttons).  The only bright side is
that there's been at least two large test suites which rely on the same
behaviour as we do here.

Disclaimers
 - Screen readers will still miss the longer help-inline when tabbing
   through the form.

 - The behaviour differs from <label for="x"> used in the rest of the form
   in that this label is not clickable.  Par for the course IMO
   (label click targets incorrectly include large white spaces;
    clicking on the field names on the left hand side will toggle
    a checkbox field...)  Someone could smooth it over in JS if they like.

 - modulebuilder still generates (invalid) labels like the ones we've
   just fixed.