Revert "timezone_menu() needs an id so the control can be labelled"
This reverts commit e8b65220180d0302a8325cc11cd34e20fa317625.

I didn't notice I was editing an upstream CodeIgniter file.
CodeIgniter/develop actually has different fix for timezone_menu(),
which is not compatible.

Note that merging the upstream fix won't do anything on its own;
each caller of timezone_menu() will have to set a new parameter,
to get an id which can then be associated with a label.