user profile: remove odd non-showing, non-functional delete button
'Site.User.Manage' does not exist, so the button can never show.
(FIXME: this doesn't cause any error or warning).

At the moment you can only view your own profile,
but you're not allowed to delete() your own user,
so this button wouldn't have worked.
And if you could view other user's profiles,
the condition for displaying the button would
still need updating to test for this condition.

Redirecting into the admin area looks slightly odd.
Also using GET for "unsafe" actions is not recommended.
And marking up what looks like a button as a HTML
anchor element is not completely ideal.

I suspect there will be an "unregister me" function in future. But that
could equally be implemented in the user.php controller, so it wouldn't
need to use a clunky cross-controller link.  (E.g. it could be a
secondary submit button on the profile form).