Consistent escaping of site.title
site.title was escaped e.g. in mobile__header.php
(and is used in at least one attribute value).

Obviously no attack vector.  The point is the config item should either
allow markup, or not.  If it doesn't allow markup, it should handle
characters like '&', '"', and '<' correctly in all cases.
(OR document it as a problem).  And it is used inside <title>,
so we don't really want to allow custom markup.