Fix invalid table markup
One duplicate <tfoot>.  Mismatches in the number of columns
(which actually show up as small glitches in the table borders,
at least in Firefox).

This leaves some garbage in the profiler that happens to be an invalid
table - it's a bug in CodeIgniter.  It's fixed upstream - note that they've
also changed the type of markup generated - someone will have to change the
profiler template anyway, and hopefully we'll get valid markup out of it.