log module: "3 Log files deleted" should be "3 log files deleted"
I don't know about the Persian; I've left that unchanged.
The position of %d appears to match lang('activity_deleted') [1],
so it should at least be in the right logical position.  The question is
whether the translator noticed that the string has a numeral prepended -
if they didn't, then the translation may not be quite right.

[1] Unless it involves some invisible control character.