BF_Model, config_file_helper: Don't strip "submit" and "func" fields
In this commit, we stop cleaning these fields out in BF_Model and
config_file_helper.  Two users of BF_Model needed modifying to clean up
their own POST data.  None of the config_file_helper users relied on
this behaviour.

It's so magic we didn't know how to document it.  So magic we forgot
about "func", which is not cleaned consistently and not used anywhere.
So why did I even notice, you ask?

I want to change the delete link inside an edit form into an secondary
submit button, because standard GET links don't benefit from
CodeIgniter's protection against CSRF.  There might be better ways to
do it, that e.g. don't send all the un-needed form data when deleting,
but I'm going ahead because it

 + works
 + has the advantage of not actually being wrong
 + is obvious what the code is doing
 + avoids me having to hack up something more complex with javascript

Given that, I need to distinguish the two submit buttons.  Using value=
doesn't work, because I need <button> to allow for the delete icon, and
IE7 doesn't send the value for <button>; it sends the element content
instead.  Grr, IE.  So I use name="delete", and distinguish the main
submit button with name="save" (matching the usual save_X() method).

This means the main submit button is no longer called "submit", we don't
clean it out of the POST data, and we choke on it.

Another entertaining effect of this redundant naming convention: if you
have a form F containing a control with the name "submit", the control
will be accessible as F.submit - making it impossible to call the DOM
method F.submit().

And hey, someone might try to use a database field called "func" someday.

"Convention over configuration" is a fine idea, but this specific magic
doesn't look like a good implementation of it.

We're not even taking advantage of this convention in the output of the
modulebuilder (which could be said to serve as a form of documentation).
Perhaps because passing _POST directly to update() would allow users to
update any field - including delete flags, dates, or id... possibly
bypassing naive form validation code.  I don't think it's a big problem
where this is currently used in the core, but it does make me feel
uncomfortable.