Enable CSRF protection
Description from config.php:

"Enables a CSRF cookie token to be set. When set to TRUE, token will be
checked on a submitted form. If you are accepting user data, it is strongly
recommended CSRF protection be enabled."

Otherwise, if you visit another site while you're logged in to Bonfire,
the other site can initiate arbitrary actions on Bonfire.