Escape form input values by fixing usage of set_value()
Double quotes and so on need to be escaped inside attribute values.
In most cases set_value() does this for us, but that doesn't work
when we bypass it :).

It looks like this also fixes a bug in these forms.
Now, if you enter an invalid value and hit submit,
you'll get _your_ value back, so you can edit it
and try again.  Previously, the value would have
been reset to the original from the database,
which could be a bit frustrating.