gravatar_link(): blind users don't need to know that Bonfire uses gravatars
The empty string is a valid value for the alt attribute.
It is similar to aria-role="presentation".

In the admin topbar, blind users should be able to read the users display
name (assuming we haven't screwed up somwhere).

If gravatar_link() is used in content, e.g. in a forum or blog, the calling
code should do the same thing: show the display name as well, which will be
very useful to everyone.