users auth: remove unecessary argument checking
If it does not make sense to call your function with the default
parameter value, then don't give it one.  If someone makes a coding
error and calls it without enough arguments, PHP will give them a
coding error.  It's easier to debug than if it silently returns FALSE.

And the empty() string / integer is not inherently more problematic than
a non-empty() value which also doesn't exist in the database, or
whatever...  OK, it's an edge case, and sometimes it would need special
handling.  But repeating this pattern in every single function is not
the answer, and none of these checks are stopping something bad from
happening.