emailer lib: implement send($d, FALSE) as expected by email test method
FALSE was the default value, so passing it didn't achieve anything.
What the test method was _trying_ to do was force the email to bypass
the queue, so we could get useful debugging information.  (For testing the
queue there's already an "insert test email" button, which passes TRUE).

Treat $queue_override as a tristate: null by default, but can be set to
either TRUE or FALSE.