template: remove magic behaviour of set('debug', ...)
This was mentioned in the old docs, but we're not using it anywhere.
If you use any of these names:

    debug|active_theme|default_theme|current_view|layout|parse_views|
    data|blocks|message|theme_paths|site_path|orig_view_path|ci

they will set class properties of Template, instead of adding a
variable to the view data.  I noticed this when trying to pass a
variable called 'debug' to the view.  The name 'data' also looks ready
to cause many man-hours of debugging fun.

There's no alternative setter for $debug, so I just made it public.
(Remember that without this change, _all_ the properties of the
Template class are effectively public.)  The other properties could
already be set as desired without using this magic.