emailer lib: process_emails(): respect enable_debug() and fix return value
There's not much point testing for CI_Session because it will always exist.
It's loaded by the users auth library, which is loaded by MY_Controller.

OTOH, it may well make sense to respect enable_debug().  This will help
make the debug code a bit more consistent in future.  I.e. we can use a
single PHP variable to hold debug messages, so we don't need to abuse the
session for it.  Being consistent about enable_debug() would make that
easier to explain and understand.

Also fix the return value to indicate success or failure.  It's what the
function comment says, and you shouldn't have to enable debugging just to
tell whether the function succeeded or not.