emailer lib: fix return value of send() and unify debugging messages
The function comment, users module, and create method of the email
settings controller think it returns boolean, but it doesn't.
Returning boolean would make good sense here.  Most callers won't
use the extra debug information - you already have to do extra
work to enable the debugging messages.  And for those that do use them,
grabbing them from a global variable is no big hardship.

(PHP does not do threading - or signal handling - and process_emails()
cannot be called recursively, so a "global variable" is not so bad).

Let's fix send() so it matches the documentation, and so the naive callers
who believed it will now notice when send() fails.

Also fix resend_activation(), which was trying to display the debug
messages, but would actually have errored out trying to display the value
of the private $debug flag.