form_dropdown(): cleanup wrong handling of $data parameter
data['name'] was cleared, but then tested for later on.  (Also, the later
test didn't even check whether $data was an array first - strings are also
possible).

Removing the name from $data, and putting it in $defaults wasn't actually
necessary.  That was the only use for the $defaults variable, so we can get
rid of it.