builder: remove variable $form_input_delimiters
I want to be able to alter this markup for radio buttons specifically.
This will allow improving the association with labels, as I did with
the other radio buttons in Bonfire.

Having a variable doesn't makes as much sense now we're using Bootstrap,
so let's hard code it.  That means I can easily edit the individual
instance for radio buttons (and use string interpolation), while leaving
the other instances unchanged.