test_permission_model: restore production tables names (issue #404)
> When I run the Bonfire Test Suite, in test_role_model, this error is
> returned:
> Table 'bonfire_dev.bf_mock_roles' doesn't exist

For me, the test immediately after the permission_model is user_model, and
that's the one that dies.  I suspect the order is non-deterministic - at
least on Unix - i.e. it's in directory traversal order.  And presumably
the tests were last run on a system where test_permission_model came last.