Modify model's update to be more flexible
Accepts either the primary_key value that was previously accepted as
the first argument, or an array to supply as a where clause. The $data
array is still the second argument, followed by a $limit value as
available on $this->db->update().

We could potentially deprecate update_where() as well.