Remove use of Template::redirect() in the admin area
This is pure cleanup, not trying to solve any functional problem.

bonfire/application contained 70 redirects overall.  30 of them were
Bonfire's Template:redirect(), as opposed to CodeIgniter's redirect().
In the admin area, Bonfire no longer uses AJAX, so they make no difference.

New code like the language module doesn't use Template:redirect() at all.
Any fine distinction that's not currently used is just going to get
bitrotted and confusing.

There are a few uses outside the admin area (and in libraries which can be
used outside it).  Leave them alone; they may be useful.  In particular,
the redirect when auth::restrict() fails, which we want for the remaining
AJAX on the email test button.