template: remove unjustified, buggy header overrides for ajax
The Content-Type here desperately needs to set the character encoding
(UTF-8).  But there's no reason to override the default Content-Type
anyway.

Similarly for caching; boo hoo users can't use F5, that's what AJAX means.
(The only one we have at the moment is POST, so it won't be cacheable
anyway).