themes: add jquery.cookie, v1.2-8-g107eb9f (currently unmerged!)
I.e. current master + https://github.com/carhartl/jquery-cookie/pull/104

A cookie function is the easiest way to get the CSRF token for AJAX
requests.  We'd like to use the jquery cookie plugin (other people have
recommended it for this exact usage).  Include it in the default theme
as well as the admin theme, for sites which use AJAX in their front end.

Unfortunately the jquery cookie plugin was reported as failing randomly in
IE.  We don't want random failures.  Especially as lots of AJAX code, like
Bonfire 0.5, doesn't bother with error handling so the failures would be
silent as well as random.  This is why we need the pull request mentioned
above.