install: fix double-wrapping of errors (invalid markup and visibly wrong)
The message() function in install_helpers.php already wraps the messages
once.  Remove the outer layer of markup.  This avoids nested <p> tags, as
well as the visual effect of doubly-nested boxes.

All the view needs to do is echo a single variable.  This is similar to
how Template::message() is used, in the main part of Bonfire.