install: Turn "unsafe" GET link into POST form button
HTTP/1.1:
   In particular, the convention has been established that the GET and
   HEAD methods SHOULD NOT have the significance of taking an action
   other than retrieval.

The usual horror story is of browsers with some form of prefetch for normal
links.  (Though this one is a minor issue, compared to the old delete
"buttons").

*Not* a security fix, despite the message I used for errors.  Installer
doesn't enable CSRF protection - it's not authenticated anyway.
(And currently it doesn't have access to the form helper, to provide the
CSRF token).