Fix (an) installer infinite redirect bug
Trying to visit /admin/ while the installer directory is present causes
an infinite redirect loop to /admin/install.  This is because index.php
generates a (non-conforming) _relative_ 302 redirect.

But we kind of wanted a relative redirect, in order to support installing
to a subdirectory.

Steal the auto-subdirectory idea from CodeIgniter's Config::__construct().

Probably no point trying to look at the config option for this.  Bonfire
installer doesn't ask you to set it in the config (only RedirectBase in
.htaccess).  If we wanted to do that, we'd have to move this somewhere
much less obvious - as a pre-controller hook, so it runs after the Config
class is loaded, but before we _potentially_ try to access a non-existent
database and die with an unhelpful error.