Fix new unassociated form labels [builder create_context]
Found with the help of http://wave.webaim.org/

`<label for="context_name">` targets the element with _id_
"context_name", not all the form elements with a _name_ of "context_name".
Add the missing id which the label was targeting.

The second label is for an entire group.  But screen readers won't see it
when tabbing through form fields, because it's not actually associated
with anything.  Use the ARIA group hack (for more comments see a2068ee7).