unit_test: fix CSRF protection error on first run
Lonnie demonstrated that if the unit test was invoked without having first
visiting Bonfire, the unit test would die.  The test assumed that the CSRF
cookie had been set by visiting some other Bonfire page.

Fix it so the unit test only tries to read the cookie on POST.  The
cookie will be set during the initial visit to the unit test (which will
be using GET).