constants: fix comment - IS_AJAX is not a security measure
Fortunately we don't use it at all internally.  But it shouldn't be
advertised as providing security, because it doesn't.  (It might prevent a
CSRF, but you need a proper method that protects normal forms as well, so
this is not to be recommended).

I've left the constant in.  Maybe someone uses it in view files for
convenience?