Template::redirect(): fix invalid DOM for AJAX & don't use JS for non-AJAX
Oops.  Invalid DHTML is still invalid, it's just harder to notice.
We shouldn't output a title element if the html is going to be inserted
into the body of another page.  (But we _should_ if the html is going to be
used as a standalone document).

We can try and fix this without changing existing code (including
out-of-tree code).  Change the response depending on whether the request
was made using AJAX.  We can serve the javascript without the title etc to
AJAX requests.  For non-AJAX requests, we can send a standard 302 redirect
and avoid the issue of valid html.

Using a 302 redirect allows writing front-end pages that work when JS is
disabled.