users auth restrict(): prevent infinite loop of Template::redirect()s
previous_page is broken in other ways (multiple tabs).  Still, if they hit
this specific bug they could get quite concerned, and it's easy enough to
suppress.

That said, if the redirect $url was passed by the caller, it's better not
to catch it here.  This will keep the loop obvious, so the caller can be
fixed.

This is distinct from an infinite loop of CodeIgniter redirect()s, in that
it doesn't use a standard HTTP redirect, where browsers would detect
excessive redirects and abort with an error message.