Remove xss_clean and strip_tags from form validation (except module builder)
(Modulebuilder could be done later if desired).

xss_clean is just broken, as mentioned in earlier commits.
Avoiding calling this kludge may even have performance benefits :).

strip_tags is not generally a good idea either, because it silently
mangles user input.

To re-iterate, the idea is that "dangerous" user input is either

 - escaped on _output_
 - rejected as invalid