users profile (only): fix redirect after login
1. Make sure you're logged out
2. Visit /users/profile and log in
3. After logging in, you should be redirected bad to /users/profile

This fixes it.  I haven't root caused it yet, but fortunately
this is an obviously correct simplification :).

I notice that the custom code here was using CI redirect().
The code in auth->restrict() is using Template::redirect() instead,
which I assume is causing the difference in behaviour here.
Perhaps this is the cause of #640