Escape the strings used in onclick attributes
Translations could include apostrophes,
or other characters that need to be escaped.

(One of the english strings already used &rsquo; and so had to be fixed.
Translators shouldn't be expected to know this kind of escaping is
necessary, if they add an apostrophe where the English string didn't
have any).