translate edit: don't automatically fill in English values for new fields
Leave them unset - it'll be easier to spot new values that way.
(And comments/whitespace are now taken from the English file).

The English fallback is moved to runtime.  Yes, it's slower, and that'll
feel unfair if you have a complete translation anyway...  As always,
criticism and patches welcome.  (Anyone want to convert us to gettext()?)