Move current_user cache out of the controller
HMVC can have multiple controllers.  In fact, there are a minimum of two:
the initial controller, and the "application object".  See `MX/Base.php`.

So the obvious place for the current_user cache is the users library.
This can then be used from other libraries.  See issue #671.

For b/c and laziness, the controller still takes a reference to the
current user.