users auth: don't support bypassing is_logged_in()
This should only make a difference when someone's password is changed.
(Clients shouldn't be able to tamper with the session, because it's
protected with a key and a checksum).  Still, it looks very very wrong.

Also it doesn't affect restrict(), because that already tests
is_logged_in().  However it would affect unguarded use of has_permission().