users auth: fix "must provide a numeric user id" on login
[issue #679]

The need for additional comments suggests this hack may be
more evil than elegant.

Still, I think this is preferable to the old situation.
Where some methods didn't test is_logged_in().  So
you had to reason about what that meant, and there
weren't any comments that explained it for you.

No idea what "best practice" should be for int / numeric strings
near the database.  Existing behaviour is preserved as much as
possible.  (But current_user->role_id is now a copy of auth->role_id(),
so that changes to an int now).