log view: try to avoid php warning when file doesn't exist
I say try because this is racy, but PHP doesn't provide any better
option.

It should be OK not to log this error, because it's only for the file
not existing.  Any more interesting errors will still cause a PHP warning
(which should at least show up in the log).  Permission errors would be an
obvious example.