index.php, install/index.php: Add master $path variable
This replaces 174b9e7e "Don't repeat '../bonfire'", to allow configuring
install/index.php in the same way as index.php.  (Subsequent commits will
fix all the paths in the installer code, to respect this configuration).

As before - I'm not sure why this wasn't sorted properly, and this commit
doesn't exactly solve the problem.  It'll only help if you know to look
in the index.php files (both of them!).

Effectively this generalizes the earlier commit to "Don't repeat '..'".
The repeated "/bonfire" segment is fairly hard to miss, and this way could
be simpler.  Especially if you're not sharing one BFPATH between several
apps.