install: fix mistaken uses of FCPATH
Many of the uses of the FCPATH constant in installer_lib were wrong -
they wanted the FCPATH used by the installed app, not the FCPATH used
by the installer app.  The installer still worked because the affected
code is either dead or non-essential.

One affected instance of FCPATH.'../' can be replaced with INSTALLPATH,
so it now respects the value of the $path variable in install/index.php.