install: fix search path for migrations belonging to custom modules
Just having 'bonfire/modules' is wrong (out of date), because this is used
to install migrations from custom _application_ modules.  (Bonfire _core_
doesn't have module-specific migrations at the moment).

I'm just copying over the locations from application/config/config.php.
I'm suspicious about '../..' in the values (it looks like an attempt to
break out of the public/ folder, but we want that to be relocatable).
However AFAICT the installer doesn't use the values, so it's ok for now.