migrations module: remove config['migrations_path']
Remove the migrations_path config option.  We also notice that the removed
comment claims the path needs to be writeable - this is not true now, so
we can remove the corresponding check from the installer.

Even before the new directory structure, it didn't make any sense to change
the migration path.  (Bonfire looked in a "core" subdirectory for it's own
migrations, so you would have had to copy that over as well).

In the new directory structure, the application has its own migrations
directory.  But the config didn't allow you to change that.  It only
allowed you to change the bonfire/migrations directory!  (The installer
uses the corresponding constructor parameter, but that doesn't mean we
need a config option, and we could easily get rid of that if we wanted to
anyway).