migrations library: installer doesn't have application_helper
(A copy of) the migrations library is used by the installer.  The installer
doesn't have application_helper (selected helpers are placed in
installer_helper).

Fortunately the applications_helper is in config/autoload.php, so
there's actually no need for the migrations library to explicitly depend on
application_helper.

Again, the aim is to make the installer slightly easier to maintain :).