migrations library: fix broken install() method
install() wasn't updated to match the new directory layout.  Also it's
redundant with the get_latest_version() method.  The extra filtering in
install() is interesting, but being inconsistent about that is obviously a
Bad Idea.

Fix get_latest_version() to use the stricter filtering, (but still use
the correct paths).  Then fix install() so it just calls
get_latest_version().