migrations library: remove set_path() method
The migrations controller used set_path() for module migrations, but that
wasn't needed, at least not any more.

The problem with setting the path is there are actually three different
paths.  (This is more obvious following the new directory structure, but
I think the same problem applied before as well).