AUTO_INCREMENT portability fix (#750)
AUTO_INCREMENT isn't portable SQL.  dbforge converts it to different types
in different databases, and they don't all work the same.

Inserting 0 or NULL into an AUTO_INCREMENT column is MySQL-specific.
It doesn't work in PostgreSQL.  Use the obvious portable approach of not
assigning a value to the AUTO_INCREMENT column.

It seems this approach obvious to most other people as well :).
This only affected one migration mentioned in issue #750,
and the creation of new roles.

There was similar-looking code in users and builder which shouldn't affect
anything, but I've tidied it away so we can be sure.