install: remove unreliable check for mod_rewrite (issue #772)
Unfortunately this check doesn't appear to be reliable.  The information
was provided by PHP's apache2handler.  I guess one explanation is that a
different handler was being used (generic cgi/fastcgi).

The removed comment may be confusing, so note that mod_rewrite is an
Apache module and not a PHP extension :).