install: avoid accidentally re-creating installed.txt (issue #761)
The installer pages aren't as paranoid / robust as they could be.
This commit just tweaks one case that bit me.

If you install, then remove installed.txt, then re-visit
install/index.php/complete (which shows a fair amount of information you
might have been interested in) - it causes installed.txt to be re-created.

Move the code that creates installed.txt so it's somewhat less likely to
be triggered.  With this tweak, you'd have to go back a second time to
the "do_install" url (the method that creates the database, among other
things).