users model: don't try to insert NULL username; this field must be NOT NULL
update() didn't try to do this.  Rip it out.

This fixes a database error when creating users with an empty username,
which is expected to work (if you've set auth.use_username=0).

(We still don't support empty usernames properly, because it stops you
editting users - issue #659).

Thanks to @artlantis for pointing this out in issue #783.