Fix: Unstyled Emailer sub-menu in Settings menu
When a submenu is built from only one module, the Contexts library was
not applying the Bootstrap classes to the submenu. The change in
build_item() actually just places most of the method's prior code into
the if (empty($menu_view)) condition. The else portion uses primarily
code adapted from the build_sub_menu() method.

More than likely there's a more efficient way to handle this, at least
from the standpoint of code reuse (if not performance).