Add basic support in installer for other DB types
Added basic support for all DB types for which we currently have
drivers, plus a placeholder if the MongoDB driver is updated in the
future.

Note the support for ODBC and PDO is very limited, and many of the
connection options for Oracle (oci8) are not supported.

None of these have been added to the supported_dbs array in the install
controller, yet.