Potential Fix for #831, incorrect application_lang.php file loads
Because the Auth library wasn't loaded early enough in the process, the
application_lang.php file was not loading the current user's language,
and right-to-left language support was not working correctly.

The only real changes in the code were to move the loading of the auth
library before the call to the parent's constructor in the
Authenticated_Controller and remove the call to
$this->set_current_user() (since this is called by the parent's
constructor).

For some reason spaces were used for all of the indentation in the file,
but now it is using tabs, which should account for the diff.