Allow suppression of newlines in Assets::image()
When using Assets::image() to generate an img tag it can be preferable
to suppress the newline output by the method (for example, if the image
is to be used inside an anchor (using anchor() or otherwise) or when
multiple images need to be aligned).

This is one of the few areas in which browsers actually notice
whitespace, so an anchor, for instance, will display some visible
extension beyond the image if there is whitespace between the opening
and closing tag.