Fix sort on System Logs
On systems that do not return the list of files in alphabetical order,
the list of logs would simply be the reverse of whatever order they were
returned. Used arsort() instead of array_reverse() to sort the list of
files in reverse alphabetical order, so the logs are listed from newest
to oldest.