Added a simple statistics reporter that reports anonymous environment statistics back to http://cibonfire.com.
The collected data is:

	- Bonfire version
	- PHP version
	- Server Type (Apache, nginx, etc)
	- DB Server Version
	- DB Client Version
	- Curl Installed?
	- Install date (just to track usage over time)
	- A server hash which is an md5 hash of: server name, server addr, and server signature

NOTE: This is not live yet, but is being prepped for the new website which will begin collecting stats. Once more stats come in we'll be able to get a more accurate idea of how many installs, what types of databases are being used, server environment, etc. and will create a display page for these stats.